requirements.txt
fastapi
uvicorn
PyMuPDF
Pillow
pytesseract
